prod düzenlemeleri.

+++ ProjectManagement/Model/SeriesCalcDailyParam.cs
@@ -7,7 +7,7 @@
         public int? StartYear { get; set; }
         public int? EndYear { get; set; }
         public int? PvCalculation { get; set; }
-        public int? PeakPower { get; set; }
+        public double? PeakPower { get; set; }
         public int? Loss { get; set; }
         public string? OutputFormat { get; set; }
         public int? UseHorizon { get; set; }
